+++ server.rb
@@ -1,16 +1,28 @@
 # encoding: UTF-8
 
+require 'optparse'
+require 'yaml'
 require 'sinatra'
 require 'cgi'
 
-require './blank'
-require './fuzzy'
+Dir["#{File.dirname(__FILE__)}/lib/*.rb"].sort.each do |path|
+  require path
+end
 
-#music_directory = 'J:/Music/'
-music_directory = '/Users/seshbaugh/Music/iTunes/iTunes Media/'
+OptionParser.new do |opts|
+  opts.on('-p', '--port [PORT]', 'Listen on the specified port.') do |port|
+    set :port, port
+  end
+end.parse!
+
+application_settings = YAML::load_file('config/settings.yml')[settings.environment]
+
+application_settings.each do |application_setting, value|
+  set application_setting, value
+end
 
 get '/' do
-  @pwd = music_directory
+  @pwd = settings.music_directory
 
   @directories = get_directories(@pwd)
 
@@ -23,7 +35,7 @@ get '/:artist' do
   if check_blacklist(params[:artist])
     status 404
   else
-    @pwd = music_directory
+    @pwd = settings.music_directory
 
     add_breadcrumb(@pwd)
 
@@ -45,7 +57,7 @@ get '/:artist/:album' do
   if check_blacklist(params[:artist]) || check_blacklist(params[:album])
     status 404
   else
-    @pwd = music_directory
+    @pwd = settings.music_directory
 
     add_breadcrumb(@pwd)
 
@@ -57,7 +69,7 @@ get '/:artist/:album' do
       if File.exists?(File.join(@pwd, params[:album]))
         @pwd = File.join(@pwd, params[:album])
 
-        @files = get_files(@pwd)
+        @files = get_files(@pwd).delete_if { |song| check_blacklist(song) }
 
         add_breadcrumb(params[:album])
 
@@ -75,7 +87,7 @@ get '/:artist/:album/:song' do
   if check_blacklist(params[:artist]) || check_blacklist(params[:album]) || check_blacklist(params[:song])
     status 404
   else
-    @pwd = music_directory
+    @pwd = settings.music_directory
 
     if File.exists?(File.join(@pwd, params[:artist]))
       @pwd = File.join(@pwd, params[:artist])
@@ -100,12 +112,12 @@ end
 not_found do
   @request_path = request.env['REQUEST_PATH'].split('/')
 
-  @pwd = music_directory
+  @pwd = settings.music_directory
 
   if @request_path.length > 2 && File.exists?(File.join(@pwd, CGI::unescape(@request_path[1])))
     @pwd = File.join(@pwd, CGI::unescape(@request_path[1]))
 
-    @artist = CGI::unescape(request_path[1])
+    @artist = CGI::unescape(@request_path[1])
 
     if @request_path.length > 3 && File.exists?(File.join(@pwd, CGI::unescape(@request_path[2])))
       @pwd = File.join(@pwd, CGI::unescape(@request_path[2]))
@@ -124,26 +136,25 @@ not_found do
     @closest = FuzzyMatch::find_closest_match CGI::unescape(@request_path.last), @files
   end
 
-  erb :'404', layout: false
+  erb :'404', :layout => false
 end
 
-
 def check_blacklist(param)
-  blacklist = ['Automatically Add to iTunes', 'Automatically Add to iTunes.localized', 'Mobile Applications', 'Album Artwork']
-
-  !blacklist.index(param).nil?
+  !settings.file_blacklist.index(param).nil?
 end
 
 def get_directories(pwd)
-  Dir.entries(pwd, { encoding: 'UTF-8' }).select { |entry| File.directory? File.join(pwd, entry) and !(entry == '.' || entry == '..') }.delete_if { |directory| check_blacklist(File.split(directory).last) }
+  Dir.entries(pwd, { :encoding => 'UTF-8' }).select { |entry| File.directory? File.join(pwd, entry) and !(entry == '.' || entry == '..') }.delete_if { |directory| check_blacklist(directory.split('/').last) }
 end
 
 def get_files(pwd)
-  mp3_files = Dir[File.join(pwd, '*.mp3')]
+  files = []
+22
+  settings.allowed_file_types.each do |file_type|
+    files += Dir[File.join(pwd, "*.#{file_type}")]
+  end
 
-  m4a_files = Dir[File.join(pwd, '*.m4a')]
-
-  (mp3_files + m4a_files).sort
+  files.sort
 end
 
 def add_breadcrumb breadcrumb
@@ -162,6 +173,6 @@ helpers do
 
     template = ('_' + template.to_s).to_sym
 
-    erb template, { layout: false }, locals
+    erb template, { :layout => false }, locals
   end
 end